+++ core/src/main/kotlin/org/sert2521/sertain/coroutines/Utils.kt
@@ -2,25 +2,20 @@ package org.sert2521.sertain.coroutines
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.cancel
+import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import org.sert2521.sertain.events.Change
 import org.sert2521.sertain.events.subscribe
 import org.sert2521.sertain.events.Event
-import org.sert2521.sertain.events.False
 import org.sert2521.sertain.events.TargetedEvent
-import org.sert2521.sertain.events.True
-import org.sert2521.sertain.events.fire
-import org.sert2521.sertain.events.subscribeBetween
 import org.sert2521.sertain.units.Chronic
+import org.sert2521.sertain.units.MetricUnit
 import org.sert2521.sertain.units.MetricValue
+import org.sert2521.sertain.units.convert
 import org.sert2521.sertain.units.from
 import org.sert2521.sertain.units.milliseconds
 import org.sert2521.sertain.units.ms
-import org.sert2521.sertain.units.s
 import kotlin.coroutines.coroutineContext
-import kotlin.properties.ReadOnlyProperty
-import kotlin.reflect.KProperty
 
 suspend fun delay(t: MetricValue<Chronic>) = delay(t.from(milliseconds).toLong())
 
@@ -53,6 +48,50 @@ suspend inline fun <T, reified E : TargetedEvent<T>> delayUntil(target: T) {
 suspend fun delayForever() {
     while (true) {
         delay(2000)
+    }
+}
+
+suspend inline fun periodic(
+        period: MetricValue<Chronic> = 20.ms,
+        delay: MetricValue<Chronic> = 0.ms,
+        time: MetricValue<Chronic> = 0.ms,
+        crossinline action: () -> Unit
+) {
+    val job = CoroutineScope(coroutineContext).launch {
+        delay(delay)
+        while (true) {
+            action()
+            delay(period)
+        }
+    }
+    if (time.value > 0) {
+        delay(time)
+        job.cancelAndJoin()
+    }
+}
+
+suspend inline fun timer(
+        period: MetricValue<Chronic> = 20.ms,
+        delay: MetricValue<Chronic> = 0.ms,
+        time: MetricValue<Chronic> = 0.ms,
+        crossinline action: (MetricValue<Chronic>) -> Unit
+) {
+    val startTime = System.currentTimeMillis()
+    periodic(period, delay, time) {
+        action((System.currentTimeMillis() - startTime).ms)
+    }
+}
+
+suspend inline fun timer(
+        period: MetricValue<Chronic> = 20.ms,
+        delay: MetricValue<Chronic> = 0.ms,
+        time: MetricValue<Chronic> = 0.ms,
+        unit: MetricUnit<Chronic>,
+        crossinline action: (Double) -> Unit
+) {
+    val startTime = System.currentTimeMillis()
+    periodic(period, delay, time) {
+        action((System.currentTimeMillis() - startTime).toDouble().convert(period.unit to unit))
     }
 }
 
@@ -92,58 +131,3 @@ class ActionGroupConfigure {
         actions.add(action)
     }
 }
-
-abstract class Observable<T>(val get: () -> T) : ReadOnlyProperty<Any?, T> {
-    val value get() = get()
-
-    var lastValue = value
-        private set
-
-    init {
-        RobotScope.launch {
-            periodic {
-                when {
-                    lastValue != value -> fire(Change(this@Observable, value))
-                }
-                lastValue = value
-            }
-        }
-    }
-
-    fun CoroutineScope.onChange(action: suspend CoroutineScope.(event: Change<T>) -> Unit) =
-            subscribe(target = this@Observable, action = action)
-
-    override fun getValue(thisRef: Any?, property: KProperty<*>) = value
-}
-
-open class ObservableValue<T>(get: () -> T) : Observable<T>(get) {
-    open operator fun invoke(configure: Observable<T>.() -> Unit) = apply(configure)
-}
-
-class ObservableBoolean(get: () -> Boolean) : Observable<Boolean>(get) {
-    init {
-        RobotScope.onChange {
-            when (it.value) {
-                true -> fire(True(this@ObservableBoolean, it.value))
-                false -> fire(False(this@ObservableBoolean, it.value))
-            }
-        }
-    }
-
-    operator fun invoke(configure: ObservableBoolean.() -> Unit) = apply(configure)
-
-    fun CoroutineScope.whenTrue(action: suspend CoroutineScope.(event: True) -> Unit) =
-            subscribe(target = this@ObservableBoolean as Observable<Boolean>, action = action)
-
-    fun CoroutineScope.whenFalse(action: suspend CoroutineScope.(event: False) -> Unit) =
-            subscribe(target = this@ObservableBoolean as Observable<Boolean>, action = action)
-
-    fun CoroutineScope.whileTrue(action: suspend CoroutineScope.(event: True) -> Unit) =
-            subscribeBetween<Observable<Boolean>, True, False>(target = this@ObservableBoolean, action = action)
-
-    fun CoroutineScope.whileFalse(action: suspend CoroutineScope.(event: False) -> Unit) =
-            subscribeBetween<Observable<Boolean>, False, True>(target = this@ObservableBoolean, action = action)
-}
-
-fun (() -> Boolean).watch(configure: ObservableBoolean.() -> Unit = {}) = ObservableBoolean(this).apply(configure)
-fun <T> (() -> T).watch(configure: ObservableValue<T>.() -> Unit = {}) = ObservableValue(this).apply(configure)
